requirements.txt
jupyter
earthengine-api
geemap
geedim
ipyleaflet
eerepr
hydra-core
wandb
piq
scipy
seaborn
torch
torchvision
python-dotenv
optuna-dashboard
google-auth
geopandas
streamlit
streamlit-flow-component
streamlit_drawable_canvas
leafmap[cesium]
keplergl
pandas
loguru
tqdm
rasterio
matplotlib
opencv-python-headless
omegaconf
Pillow
xarray
typer
optuna
